add comment popup
